iteratortesteado

+++ src/general/Validator.java
@@ -70,17 +70,17 @@ public class Validator {
 	public static boolean hasWord(String word, int[] letters) {
 		//TODO se podria mejorar el codigo y ser mas eficiente, ya ponerlas en el tablero
 		int i=0;
-		for(i=0; i < word.length() && letters[(int)(word.charAt(i)-'A')] >=1; i++){
+		for(i=0; i < word.length() && letters[(word.charAt(i)-'A')] >=1; i++){
 			letters[(int)(word.charAt(i)-'A')]--;
 		}
 		if(i == word.length()){
 			for(i=0; i < word.length() ; i++){
-				letters[(int)(word.charAt(i)-'A')]++;
+				letters[(word.charAt(i)-'A')]++;
 			}
 			return true;
 		}
 		for(; i >= 0 ; i--){
-			letters[(int)(word.charAt(i)-'A')]++;
+			letters[(word.charAt(i)-'A')]++;
 		}
 		return false;
 	}
